FIX: extract_at_coordinate (mrconvert) takes now Int and not FLoat as input

+++ cmtklib/interfaces/mrtrix3.py
@@ -319,7 +319,7 @@ class MRConvertInputSpec(CommandLineInputSpec):
     out_filename = File(genfile=True, argstr='%s', position=-1, desc='Output filename')
     extract_at_axis = traits.Enum(1,2,3, argstr='-coord %s', position=1,
                            desc='"Extract data only at the coordinates specified. This option specifies the Axis. Must be used in conjunction with extract_at_coordinate.')
-    extract_at_coordinate = traits.List(traits.Float, argstr='%s', sep=',', position=2, minlen=1, maxlen=3,
+    extract_at_coordinate = traits.List(traits.Int, argstr='%s', sep=',', position=2, minlen=1, maxlen=3,
         desc='"Extract data only at the coordinates specified. This option specifies the coordinates. Must be used in conjunction with extract_at_axis. Three comma-separated numbers giving the size of each voxel in mm.')
     voxel_dims = traits.List(traits.Float, argstr='-vox %s', sep=',',
         position=3, minlen=3, maxlen=3,
